Removed all references to view
Closes #40
