+++ include/experimental/__p1673_bits/conjugate_transposed.hpp
@@ -40,75 +40,33 @@
 //@HEADER
 */
 
-#ifndef LINALG_INCLUDE_EXPERIMENTAL___P1673_BITS_PACKED_LAYOUT_VIEW_HPP_
-#define LINALG_INCLUDE_EXPERIMENTAL___P1673_BITS_PACKED_LAYOUT_VIEW_HPP_
+#ifndef LINALG_INCLUDE_EXPERIMENTAL___P1673_BITS_CONJUGATE_TRANSPOSED_HPP_
+#define LINALG_INCLUDE_EXPERIMENTAL___P1673_BITS_CONJUGATE_TRANSPOSED_HPP_
 
-#include <experimental/mdspan>
-#include "layout_triangle.hpp"
+#include "conjugated.hpp"
+#include "transposed.hpp"
 
 namespace std {
 namespace experimental {
 inline namespace __p1673_version_0 {
 namespace linalg {
 
-// TODO declarations need extents-see-returns-below defined
+template<class EltType, class Extents, class Layout, class Accessor>
+basic_mdspan<EltType, Extents, layout_transpose<Layout>,
+             accessor_conjugate<Accessor, EltType>>
+conjugate_transposed(
+  basic_mdspan<EltType, Extents, Layout, Accessor> a)
+{
+  return conjugated(transposed(a));
+}
 
+// FIXME Must fill in see-below.
 #if 0
-template<class EltType,
-         class Extents,
-         class Layout,
-         class Accessor,
-         class Triangle,
-         class StorageOrder>
-constexpr basic_mdspan<EltType,
-  <i>extents-see-returns-below</i>,
-  layout_blas_packed<
-    Triangle,
-    StorageOrder>,
-  Accessor>
-packed_view(
-  const basic_mdspan<EltType, Extents, Layout, Accessor>& m,
-  typename basic_mdspan<EltType, Extents, Layout, Accessor>::index_type num_rows,
-  Triangle,
-  StorageOrder);
-
-         template<class EltType,
-         class Extents,
-         class Layout,
-         class Accessor,
-         class Triangle,
-         class DiagonalStorage,
-         class StorageOrder>
-constexpr basic_mdspan<const EltType,
-  <i>extents-see-returns-below</i>,
-  layout_blas_packed<
-    Triangle,
-    StorageOrder>,
-  Accessor>
-packed_view(
-  const basic_mdarray<EltType, Extents, Layout, Accessor>& m,
-  typename basic_mdarray<EltType, Extents, Layout, Accessor>::index_type num_rows,
-  Triangle,
-  StorageOrder);
-
-template<class EltType,
-         class Extents,
-         class Layout,
-         class Accessor,
-         class Triangle,
-         class DiagonalStorage,
-         class StorageOrder>
-constexpr basic_mdspan<EltType,
-  <i>extents-see-returns-below</i>,
-  layout_blas_triangular_packed<
-    Triangle,
-    StorageOrder>,
-  Accessor>
-packed_view(
-  basic_mdarray<EltType, Extents, Layout, Accessor>& m,
-  typename basic_mdarray<EltType, Extents, Layout, Accessor>::index_type num_rows,
-  Triangle,
-  StorageOrder);
+template<class EltType, class Extents, class Layout, class Accessor>
+basic_mdspan<EltType, Extents, layout_transpose<Layout>,
+             <i>see-below</i> >
+conjugate_transposed(
+  const basic_mdarray<EltType, Extents, Layout, Accessor>& a)
 #endif // 0
 
 } // end namespace linalg
@@ -116,4 +74,4 @@ packed_view(
 } // end namespace experimental
 } // end namespace std
 
-#endif //LINALG_INCLUDE_EXPERIMENTAL___P1673_BITS_PACKED_LAYOUT_VIEW_HPP_
+#endif //LINALG_INCLUDE_EXPERIMENTAL___P1673_BITS_CONJUGATE_TRANSPOSED_HPP_
