+++ include/experimental/linalg
@@ -48,11 +48,11 @@
 #include "__p1673_bits/layout_blas_general.hpp"
 #include "__p1673_bits/layout_tags.hpp"
 #include "__p1673_bits/layout_triangle.hpp"
-#include "__p1673_bits/packed_layout_view.hpp"
-#include "__p1673_bits/scaled_view.hpp"
-#include "__p1673_bits/conjugate_view.hpp"
-#include "__p1673_bits/transpose_view.hpp"
-#include "__p1673_bits/conjugate_transpose_view.hpp"
+#include "__p1673_bits/packed_layout.hpp"
+#include "__p1673_bits/scaled.hpp"
+#include "__p1673_bits/conjugated.hpp"
+#include "__p1673_bits/transposed.hpp"
+#include "__p1673_bits/conjugate_transposed.hpp"
 #include "__p1673_bits/blas1_givens.hpp"
 #include "__p1673_bits/blas1_linalg_swap.hpp"
 #include "__p1673_bits/blas1_scale.hpp"
