+++ tests/CMakeLists.txt
@@ -8,8 +8,8 @@ endmacro()
 find_package(GTest REQUIRED)
 
 linalg_add_test(add)
-linalg_add_test(conjugate_transpose_view)
-linalg_add_test(conjugate_view)
+linalg_add_test(conjugate_transposed)
+linalg_add_test(conjugated)
 linalg_add_test(copy)
 linalg_add_test(dot)
 linalg_add_test(gemm)
@@ -18,6 +18,6 @@ linalg_add_test(givens)
 linalg_add_test(iterator)
 linalg_add_test(norm2)
 linalg_add_test(scale)
-linalg_add_test(scaled_view)
+linalg_add_test(scaled)
 linalg_add_test(swap)
-linalg_add_test(transpose_view)
+linalg_add_test(transposed)
